requirements.txt
numpy
omegaconf
hydra-core
wandb
tqdm
matplotlib